share common email schema
