+++ packages/common/src/utils/index.ts
@@ -36,3 +36,4 @@ export { default as dayjs } from './dayjs'
 export { Dayjs } from './dayjs'
 export * from './promiseWithTimeout'
 export * from './handleReservedStatus'
+export * from './email'
