+++ packages/web/src/common/store/pages/signon/sagas.js
@@ -19,7 +19,8 @@ import {
   confirmerActions,
   confirmTransaction,
   IntKeys,
-  parseHandleReservedStatusFromSocial
+  parseHandleReservedStatusFromSocial,
+  isValidEmailString
 } from '@audius/common'
 import { push as pushRoute } from 'connected-react-router'
 import { isEmpty } from 'lodash'
@@ -43,7 +44,6 @@ import { retrieveCollections } from 'common/store/cache/collections/utils'
 import { fetchUserByHandle, fetchUsers } from 'common/store/cache/users/sagas'
 import { UiErrorCode } from 'store/errors/actions'
 import { setHasRequestedBrowserPermission } from 'utils/browserNotifications'
-import { isValidEmailString } from 'utils/email'
 import { restrictedHandles } from 'utils/restrictedHandles'
 import { FEED_PAGE, SIGN_IN_PAGE, SIGN_UP_PAGE } from 'utils/route'
 import { waitForRead, waitForWrite } from 'utils/sagaHelpers'
